update backend to send total grants matching filters
