+++ packages/client/package.json
@@ -40,6 +40,6 @@
     "geckodriver": "^1.20.0",
     "node-sass": "^4.12.0",
     "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.11"
+    "vue-template-compiler": "^2.6.14"
   }
 }
